requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.7.0
gunicorn==21.2.0
flask-wtf==1.2.1
email-validator==2.1.0
flask-login==0.6.2
flask-bcrypt==1.0.1
itsdangerous==2.1.2
flask-sqlalchemy==3.1.1
pyopenssl==23.2.0
authlib==1.2.1
requests==2.31.0
